# motion controller/C#/MotionController/MotionController Sensor API/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Internal MotionController Sensor API",
    "description": "API interface for all internal sensor operations.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/device/sessions/{sessionId}/orientation": {
      "get": {
        "tags": [
          "DeviceSession"
        ],
        "summary": "Get a Device Session orientation by Session Id",
        "operationId": "GetDeviceSessionOrientationAsync",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceSessionOrientation"
                  }
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatabaseModel": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false
      },
...
      "DeviceSessionOrientation": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DatabaseModel"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "id": {
                "type": "integer",
                "format": "int32"
              },
              "deviceSessionId": {
                "type": "integer",
                "format": "int32"
              },
              "rollDegrees": {
                "type": "number",
                "format": "float"
              },
              "pitchDegrees": {
                "type": "number",
                "format": "float"
              },
              "yawDegrees": {
                "type": "number",
                "format": "float"
              },
              "rollRadians": {
                "type": "number",
                "format": "float"
              },
              "pitchRadians": {
                "type": "number",
                "format": "float"
              },
              "yawRadians": {
                "type": "number",
                "format": "float"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "created": {
                "type": "string",
                "format": "date-time"
              },
              "modified": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
...
    }
  }
}